Fix critical database error: green_light_time type mismatch
- Cast green_light_time to ::bigint in UPDATE query to prevent PostgreSQL type error
- Add migration logic to detect and convert green_light_time column if wrong type
- Handle case where column might be TIMESTAMPTZ instead of BIGINT

+++ backend/src/models/Match.ts
@@ -133,12 +133,12 @@ export class MatchModel {
   // HTTP Polling Methods
 
   /**
-   * Set green light time (server-picked random delay)
+   * Set green light time (server-picked random delay) as milliseconds since epoch
    */
   static async setGreenLightTime(matchId: string, greenLightTime: number): Promise<void> {
     await pool.query(
       `UPDATE matches 
-       SET green_light_time = $1, updated_at = NOW()
+       SET green_light_time = $1::bigint, updated_at = NOW()
        WHERE match_id = $2`,
       [greenLightTime, matchId]
     );
